Move Multiple Provider vCPE related classes into 'mp' package.
BGP model in MultipleProviderTemplate is not build any more (it is left as it comes from the user).
To build it, it was using SP template dependent classes whose structure makes no sense in MP template.
If MP template requires BGP, it's own way to configure it must be implemented.