Configure BGP with ip addresses from user input.
This patch causes BGB configuration for SP-VCPE template to be updated with user provided IP addresses.

To do so, logical interfaces in core router are sent to OpenNaaS from the GUI (within create method),
and afterwards, model used to configure BGP is updated with data of recevived core router logical interfaces.

With the values correclty updated in the model, configuring BGP from the model requires no change.