Resolve list of vCPE before any other call into the controller.
List of vCPE is resolved before other calls because if following
methods fail, Model already contains the list and menu can be correctly displayed.

In methods requiring an update of the list, the list is populated inside a finally clause,
enforcing the list is populated. This approach causes these methods to rise an exception when
populating the list fails.