Display buttons in sp logicalForm.jsp after a call to updateVRRPIP.
Buttons disappear when updating VRRP IP address from edit page.
Either there is an error in opennaas or not.
The user can NOT edit anything in the page as ALL buttons have disappeared
Missing buttons are: Switch VRRP priority, Update VRRP IP and
global update, too.

Buttons disappear because action is not set to 'update', and sp logicalForm.jsp
only displays buttons when action is 'update'.
Adding the action attribute with 'update' value solves the issue.
This is the way it's done in switchVRRPPriority which behaves as expected.

This patch adds action attribute with 'update' value in calls to updateVRRPIP.