GIM implementation and APC PDU driver as delivered by TSSG.
The current driver is configured for the PDU in Dublin and works fine here,
reporting Current, Voltage, Power and Energy at the PDU level.
It also turns on and off the specified outlet and reports its status.
This has been validated against the values/statuses reported through the PDUs web front-end.

Please note though that the PDU does not seem to support outlet-level reporting of current, power, etc.
This would need to be taken into consideration if reporting the metrics for an individual router/device.