Create VCPE ASYNC.
VCPENetworkManager.create(model) is now asynchronous.

VCPENetworkManager.create(model) now creates a VCPE resource, starts an other thread to make the build and returns the resourceId.
VCPENetworkManager.hasFinishedBuild(resourceId) tells if the build has finished or not. It can be used to poll for build to finish.
VCPENetworkManager.getBuildResult(resourceId) can be used to retrieve the building task result when it has finished (either ok or exception).

GUI calls three methods above in VCPENetworkBO.create(model) method.
This method performs the polling for build to complete, behaving the same way it did before. It remains synchronous.