Create openflowswitch.driver.floodlight bundle with FloodlightProtocolSession.
FloodlightProtocolSession instantiates a FloodlightClient, similarly than OpenerProtocolSession.
The client is empty by now (not yet functional).

FloodlightProtocolSession requires a ProtocolSessionContext with the switchId, as well as the ip address of the floodlight instance.
SwitchId is stored in the context as a parameter named protocol.floodlight.switchid.
The switchId should be of the form XX:XX:XX:XX:XX:XX:XX:XX (i.e. 00:00:00:00:00:00:00:01),
and should match the identifier floodlight uses for a registered switch.