Add JSON serialization in floodlight client.
FloodlightProtocolSession instantiates a client with anotated IFloodlightStaticFlowPusherClient interface.
The client uses CustomJSONProvider to serialize/deserialize FloodlightOFFlow data structures.

CustomJSONProvider contains a serializer and a deserializer for FloodlightOFFlow and application/json MediaType.
The mapping between model classes and serializers is configured in CustomJSONProvider.
This way, model classes are not linked to serializers.

A tests checks CustomJSONProvider serializes and deserializes as expected by floodlight.