Remove unnecesary features from platform.
This commit removes test related features from the platform.

This commit causes platform to build successfully.
It was failing because itest.helpers feature is build after the platform, but platform was requiring it.
With itest.helpers feature already being loaded in itests requiring it, it is no longer required for the platform to build.