Remove ehcache dependency in opennaas.core.security.
Ehcache 1.6 must be removed in order to be able to use springframework.contecxt.support in an environment with CXF.
The dependency is removed by using a non caching cache NullAclCache.

This commit also deletes ehcache from the platform, as it is no longer required once opennaas.core.security is not using it.