Fix error in tests regarding ToS value.
ToS value in FlowRequest is the decimal representation of the whole ToS byte.
Due to 7th and 8th bits being reserved, ToS value SHOULD be multiple of 4.

However, ToS value in FloodlightOFMatch is the decimal representation of only
the 6 non-reserved bits of the ToS byte.

Tests where using 1 as FlowRequest ToS value, which does not correspond to a valid one.
Additionally, tests were assuming tos values in FlowRequest and in generated FloodlightOFMatch shoud be equals, which is not the case.

More information regarding the confusion can be found here:
http://www.tucny.com/Home/dscp-tos

This patch documents methods regarding ToS in FlowRequest to clarify its meaning and expected values,
and updates test to behave as desired.