OFNetwork resource with a capability for flow provisioning.
OFNetwork resource represents an OpenFlow domain.
OFProvisioningNetworkCapability has methods to allocate, deallocate and replace flows in the domain.
It is also able to retrieve allocated flows, although current implementation is not in sync with underlying resources.

Implementation provided by internal driver delegates the allocation of each individual flow to IOpenflowForwardingCapability.