Remove Apache ServiceMix :: Naming (1.6.1) bundle from platform.
PROBLEM EXPLANATION:
Removed bundle [1] interferes with Apache Aries JNDI Core [2].
[1] mvn:org.apache.servicemix.naming/org.apache.servicemix.naming/1.6.1
[2] mvn:org.apache.aries.jndi/org.apache.aries.jndi.core/0.3

They both try to publish a naming service with following interface javax.naming.spi.InitialContextFactoryBuilder,
and the behaviour differs depending on which one is loaded first (race condition).

When [1] is loaded first, [2] publishes a JREInitialContextFactoryBuilder instead of OSGIInitialContextFactoryBuilder, but loads normally. In this case osgi services use [1].
When [2] is loaded first, [1] fails to load. In this case osgi services use [2].

If [1] is present, it only accepts services of the form osgi:services/<interface>.
If [2] is present, it accepts services of following forms:
- aries:services
- osgi:service
- osgi:servicelist

Notice that these forms are exlusive, resulting in services correctly published in one not being binded in the other.

[1] is being introduced in platform by following dependency tree:
Feature tree
servicemix-shared 2013.01
    jbi 1.6.1
       naming 1.6.1
       \ mvn:org.apache.servicemix.naming/org.apache.servicemix.naming/1.6.1

APPLIED SOLUTION:
Override features in the dependency tree above to exclude naming.
This way we ensure used naming service is the one provided by Apache Aries JNDI Core [2].