Refactor ShowInterfacesCommand in ChassisCapability.
It now uses ModelHelper to retrieve interfaces from the model.
It now uses getResourceFromFriendlyName() in GenerikKarafCommand to get the resource the command applies to.
Printing code is now inside printing methods, not inside doExecute().

Functionality has not changed and the command prints exactly the same it was printing before this patch.