Unify ipv4 mask format in REST API.
ipv4 mask uses its short form.
The short form is the prefix lenght in bits. e.g. 255.255.255.0 is expressed as /24.

The short form is the format expected in user inputs.
It is also the format used to display ipv4 masks to the user.

The long form is used in the model, but never exposed to the user, through the REST API.