Specify resource name using the CLI.
This patch introduces the feature to specify the name of a resource using the command line.
In order to do it, "-n" option can be used, just as follows:

OpenNaaS>resource:create -n coolName /path/to/a/junosRouter.descriptor
[INFO] Loaded descriptor at file:////path/to/a/junosRouter.descriptor
[INFO] Created resource router:coolName
[INFO] Created 1 resource/s of 1

By using this feature, the same resource descriptor may be used to create several similar resources (providing a different name for each one).
Where similar means that they would be of same type and have the same capabilities.
This is usefull in organizations having several instances of the same device model.

Trying to reuse an already existing name, causes an error being displayed:

OpenNaaS>resource:create -n coolName /path/to/a/junosRouter.descriptor
[INFO] Loaded descriptor at file:////path/to/a/junosRouter.descriptor
[INFO] Created resource router:coolName
[INFO] Created 1 resource/s of 1
OpenNaaS>resource:create -n coolName /path/to/a/junosRouter.descriptor
[INFO] Loaded descriptor at file:////path/to/a/junosRouter.descriptor
[ERROR] There is already a resource in this respository with this name: coolName
[INFO] Created 1 resource/s of 1