Nullsafe implementation for method to find matching routes for a specific request.
RouteSelectionMatches method is now used.
This method compares two RouteSelectionInput objects and determines whether a candidate matches requested requeriments.
Candidate is said to match in case candidate specifications are equal or less restrictive than the ones requested.

The use of StringUtils.isEmpty(string) and the fact of only accessing methods in string when isEmpty returned false, makes the method null safe.
At least for null values in candidate attributes, which sere the cause of the reported bug (OPENNAAS-1358).