Avoid error trace in console related with NCLMonitoringCapability WS.
This patch avoids publishing NCLMonitoringCapability as WS, which was causing an error.
It keeps publishing it as an OSGi service, but not as WS any more.

The error was caused by INCLMonitoringCapability not having any declared method to export.
If this interface is to grow in the future, it should then be published as the rest.

Fixes issue OPENNAAS-1357.