CreateAggregatedInterface implementation in junos driver.
CreateAggregatedInterfaceAction is responsible for configuring given AggregatedLogicalPort.
It uses two velocity templates:
- CREATION_TEMPLATE for configuring the aggregated interface
- ADD_TO_AGGREGATED_TEMPLATE for configuring each interface that is part of the aggregation.

CreateAggregatedInterfaceAction will fail if it is given an AggregatedLogicalPort whose name is already used by an existing aggregated interface in the junos config.
In particular, it is worth notice that launching CreateAggregatedInterface twice with same parameter will cause the second action to fail.