RemoveAggregatedInterface implementation in junos driver.
RemoveAggregatedInterfaceAction is responsible for removing the interface.
This action makes use of two velocity templates:
- DELETION_TEMPLATE which deletes the aggregated interface from the device configuration.
- REMOVE_FROM_AGGREGATED_TEMPLATE which deletes the association between existing interfaces and the removed aggregated one.

RemoveAggregatedInterfaceAction will fail if there is no AggregatedInterface with given name in the model.
This requirement is related to the action imlpementation, not capability defined.
There is need to retrieve the names of interfaces given aggregation includes, in order to remove an association in the junos configuration.

Above requirement limitates the usage of this action. Queue must be executed between a create and a remove. Otherwise remove will fail.