Content Provisioning module.
Blueprint instantiates a ContentProvisioning bean and publishes it as an OSGi service using DOSGi properties,
for it being published as a web service using REST.

Instantiated ContentProvisioning bean gets injected a reference to the IResourceManager.
It also gets injected the name of the switch to use.
The switch name is taken from a configuration file in the platform called "org.opennaas.extensions.contentprovisioning.cfg",
specifically from the property named "org.opennaas.contentprovisioning.switchName".