Avoid blueprint container reloading.
This patch sets update-strategy="none" in all blueprint property-holders.

Using update-strategy="none" is a workaround for the issue of
having the blueprint container constantly going on and off when using update-strategy="reload",
as explained here:
https://mail-archives.apache.org/mod_mbox/karaf-user/201404.mbox/%3C535A2DF6.5080400%40i2cat.net%3E

NOTE: using update-strategy="none" makes opennaas unable to react to changes made to properties in hot.
Property changes will be applied after a reboot. Thus, it forces the user to reboot opennaas to apply changes.