Add configurable vlan to circuits match.
Add the option to set a vlan that will be used in the match for all circuits stablished by NCLProvisioner.

NCLProvisioner reads the option from the file org.ofertie.ncl.network.cfg
If a VLAN different than -1 is specified, it will be introduced in NCLProvisionerCapability requests.

NCLProvisionerCapability adds the VLAN (if any) to created circuits match.

IMPORTANT NOTE:
This patch is specific for a demo, and should never reach a standard release.
It does not respect modularity because it breaks component responsabilities.